+++ STM32CubeIDE/Middlewares/Robot/kinematics/kinematics.h
@@ -14,10 +14,10 @@ enum base {
 // rad/min
 struct rpm
 {
-    int motor1;
-    int motor2;
-    int motor3;
-    int motor4;
+    float motor1;
+    float motor2;
+    float motor3;
+    float motor4;
 };
 
 // m/s
